Fixed error in tilt.
Tilt needs to adjust the direction of tilting for the hyperplanes it rolls, and
all this entails is subtracting one from the direction if it's greater than the
axis you're tilting along.

Unfortunately, tilt was doing this in a loop so the tilt would get completely
screwed up.  Took this out of the loop so that the adjustment is only done once
as it should be.