Changed the order of MPI ranks in a logical machine cuboid to be ordered in the same ordering used for compaction. It helps the cuboid of MPI ranks in the logical cuboid kept in the physical grid.
Before this change, the MPI ranks in a logical machine are sorted in
row-order and then they are mapped
to the coords of physical nodes reordered by each compaction scheme.

This is not reasonable. So, now, the MPI ranks in a logical box are
reordered by using their coords in the same method as the compaction of
physical grid and then they are mapped to the physical grid.

In addition, when the recursive splitting is used, the MPI ranks in the
logical box are also reordered using the same direction vector for the
compaction of the physical grid.

e.g.) If nodes in a physical grid are split from y->x->z direction,
    MPI ranks in a logical box are also split from y->x->z direction.

This is just to keep neighboring MPI ranks grouped in the physical
nodes.