Implemented a function to estimate the shape of logical torus for better mapping of rubik into physical torus.
It uses factorization to estimate the shape of physical torus.

If the number of PE is 4096 and the shape of allocation is 8x4x5x32,

   the estimation is done as following.

number of gemini = 128 = 4096/32

factorisation of 128 = 2^7

increment each axis in decreasing order of the dimensions of the real allocation by multiplying each factor of
the number of gemini.
start from 1 x 1 x 1

round 1 -> 2 x 1 x 1
round 2 -> 2 x 1 x 2
round 3 -> 2 x 2 x 2
round 4 -> 4 x 2 x 2
round 5 -> 4 x 2 x 4
round 6 -> 4 x 4 x 4-> now the y axis is equal to the dimension of real
allocation, so it is not incremented any more from this round.
round 7 -> 8 x 4 x 4 -> The estimation is ended here It is equal to the
number of gemini.