Modified decide_direction_vector() to place nodes in xz planes first.
This is because of asymmetrical bandwidth of links on BlueWaters(y-links
        are slower than others)

Now, recursive and row-ordering place MPI ranks in xz planes first.

The difference between them is how to place nodes in xz planes.

Row-ordering orders MPI ranks from y axis, x axis and z axis.

So in this row-ordering, holes can be concentrated in certain axes and
may not be optimal for irregular cuboids.

In the recursive ordering, the allocation is split into xz planes first
and place nodes in blocks on xz planes split by subsqeuent partionings.

It makes holes scattered in many blocks so the congestion due to holes
may be distributed.

I need to verify this assumption with real runs of MILC on BlueWaters.